CS50w/capstone/requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
Django==4.2.7
django-picklefield==3.4.0
django-q2==1.9.0
google-ai-generativelanguage==0.4.0
google-api-core==2.28.1
google-auth==2.45.0
google-generativeai==0.3.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.62.3
idna==3.11
proto-plus==1.27.0
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.2
